backend/requirements.txt
Django>=4.2,<5.0
DjangoRESTFramework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.2,<11.0
